Update 3: Task #2 'technically' completed
At this point, the Mindex coding challenge is technically "completed",
as the MVP has been satisfied based on the challenge's given taska &
requirements. However, proper documentation, code cleanup, formatting,
and additional testing/test writing has not been completed to ensure the
code is fully-functioning and complete. Future commits will be mostly
dedicated to completing these particular features.

*Note about Task #2:

This task specified that the new Compensation type and its REST
endpoints "should persist and query the Compensation from the
persistence layer.". Because the coding challenge's "What's Provided"
section of instructions stated that "While the application runs, the
data may be accessed and mutated in the database without impacting the
snapshot", I made the general assumption that "persistence" meant that
the data for Compensations should and could be written to and accessed
repeatedly *during the runtime of the application*, not outside or
between separate runtimes.

TODO:
  - Clean up code syntax & code that is definitively too unoptomized (if
    possible).
  - Write extensive JavaDocs documentation for each class, method, and
    field.
  - Go through all classes & ensure code is properly formatted across
    the board.
  - Write Tests and/or Test Classes in order to make sure the currently
    written code is truly fully functional.